Allow saving cache data to a temp directory
Adds a 'tempdir_cache' boolean option to the Market class which allows
the cached market data to be saved to the system temp directory rather
than the current working directory. This avoids polluting the working
directory with a short-term file that the user likely doesn't need to
be able to directly access themselves.

By default the tempdir is not used (for backwards compatibility) and
has to be explicitly requested by passing 'tempdir_cache = True' when
instantiating a Market object.